Updated about me section - Completely redishigned the about me section - Added interests, and what im currently working on - Made it more personal
