+++ src/lib/interests.js
@@ -1,0 +1,39 @@
+export const interests = [
+    { label: "Full-Stack Development", emoji: "💻" },
+    { label: "IoT & Hardware", emoji: "🔌" },
+    { label: "AI & Machine Learning", emoji: "🤖" },
+    { label: "UI/UX Design", emoji: "🎨" },
+]
+
+export const currentWork = [
+    {
+        name: "DIY Macropad",
+        description:
+            "Designing and building a custom mechanical macropad using Arduino Pro Micro and QMK firmware.",
+        emoji: "⌨️"
+    },
+    {
+        name: "Learning React/Next.js",
+        description:
+            "Deepening my knowledge of frontend frameworks to build dynamic, scalable web apps.",
+        emoji: "📚"
+    },
+    {
+        name: "LeetCode Practice",
+        description:
+            "Sharpening problem-solving and algorithmic thinking through daily coding challenges.",
+        emoji: "🧑‍💻"
+    },
+    {
+        name: "A-Level Revision",
+        description:
+            "Preparing for upcoming A-Level exams with a focus on Computer Science and Maths.",
+        emoji: "📖"
+    },
+    {
+        name: "Personal Blog",
+        description:
+            "Documenting my learning programming journey through LinkedIn.",
+        emoji: "✍️"
+    }
+];
