+++ src/pages/Home.jsx
@@ -14,7 +14,7 @@ export default function Home() {
         <Hero />
         <MyModel />
       </section>
-      <section id='about' className='pt-32'>
+      <section id='about' className='pt-32 '>
         <About />
       </section>
       <section id='projects' className='pt-16 max-w-full'>
